+++ src/common/url.h
@@ -13,7 +13,7 @@ extern void *url_tree;
 void url_clear (void);
 void url_save (const char *fname, const char *mode, gboolean fullpath);
 void url_autosave (void);
-int url_check_word (char *word, int len);
-void url_check_line (char *buf, int len);
+int url_check_word (char *word, size_t len);
+void url_check_line (char *buf, size_t len);
 
 #endif
